Added Copy Button for the Voice ID to Voice Lab Dashboard
When using the Voice Lab in the editor Dashboard, the voice model id is shown, but there is no way to copy it to the clipboard.

This meant I was manually typing it into some of my scriptable objects. I added a "Copy" button which copies the voice id into the user's clipboard, so they can paste it wherever they need to.